+++ src/main/java/com/mapu/global/jwt/application/JwtService.java
@@ -3,6 +3,7 @@ package com.mapu.global.jwt.application;
 import com.mapu.domain.user.domain.UserRole;
 import com.mapu.global.jwt.JwtUtil;
 import com.mapu.global.jwt.dao.JwtRedisRepository;
+import com.mapu.global.jwt.dto.AccessTokenResponseDto;
 import com.mapu.global.jwt.dto.JwtUserDto;
 import com.mapu.global.jwt.exception.JwtException;
 import com.mapu.global.jwt.exception.errorcode.JwtExceptionErrorCode;
@@ -66,15 +67,15 @@ public class JwtService {
         return new AccessTokenResponseDto(accessToken);
     }
 
-    public Cookie rotateRefreshToken(String refresh) {
-        verifyRefreshToken(refresh);
-        JwtUserDto jwtUserDto = getUserDtoFromToken(refresh, JwtUtil.REFRESH);
-        return jwtUtil.rotateRefreshJwtCookie(jwtUserDto, refresh);
-    }
-
-
     public void deleteRefreshJwt(String refresh) {
         verifyRefreshToken(refresh);
         jwtRedisRepository.deleteById(refresh);
     }
+
+    public Cookie rotateRefreshToken(String refresh) {
+        deleteRefreshJwt(refresh);
+        JwtUserDto jwtUserDto = getUserDtoFromToken(refresh, JwtUtil.REFRESH);
+        return jwtUtil.createRefreshJwtCookie(jwtUserDto);
+    }
+
 }
